Separate ml_worker and mongo functions
